pyproject.toml
[tool.poetry]
name = "vrm-lstm"
version = "0.1.0"
description = ""
authors = ["Ezequiel Schneider"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.1"
pandas = "^2.2.3"
numpy = "^2.2.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
statsmodels = "^0.14.4"
ipywidgets = "^8.1.5"
tabulate = "^0.9.0"
pywavelets = "^1.8.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
notebook = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
